+++ tests/test_validation_framework/test_metric_calculators.py
@@ -38,6 +38,7 @@ import shutil
 
 from pytesmo.metrics import with_analytical_ci, with_bootstrapped_ci, pairwise
 from pytesmo.validation_framework.validation import Validation
+import pytesmo.validation_framework.error_handling as eh
 from pytesmo.validation_framework.metric_calculators import (
     MetadataMetrics,
     BasicMetrics,
@@ -60,6 +61,9 @@ from pytesmo.validation_framework.temporal_matchers import (
 )
 from pytesmo.validation_framework.results_manager import netcdf_results_manager
 import pytesmo.metrics as metrics
+
+
+from .utils import DummyReader
 
 
 def make_some_data():
@@ -94,7 +98,7 @@ def test_MetadataMetrics_calculator():
     metriccalc = MetadataMetrics(other_name="k1")
     res = metriccalc.calc_metrics(data, gpi_info=(0, 0, 0))
 
-    assert sorted(list(res.keys())) == sorted(["gpi", "lon", "lat"])
+    assert sorted(list(res.keys())) == sorted(["gpi", "lon", "lat", "status"])
 
     metadata_dict_template = {
         "network": np.array(["None"], dtype="U256"),
@@ -591,12 +595,12 @@ def test_RollingMetrics():
 # Tests for new QA4SM metrics calculators
 
 
-class DummyReader:
-    def __init__(self, df, name):
-        self.data = pd.DataFrame(df[name])
+# class DummyReader:
+#     def __init__(self, df, name):
+#         self.data = pd.DataFrame(df[name])
 
-    def read(self, *args, **kwargs):
-        return self.data
+#     def read(self, *args, **kwargs):
+#         return self.data
 
 
 def make_datasets(df):
@@ -637,7 +641,7 @@ def testdata_known_results():
             "mse_bias": 4,
             "mse_var": 0,
             "urmsd": 0,
-            "gpi": 1,
+            "gpi": 0,
             "lon": 1,
             "lat": 1,
         },
@@ -653,7 +657,7 @@ def testdata_known_results():
             "mse_bias": 16,
             "mse_var": 0,
             "urmsd": 0,
-            "gpi": 1,
+            "gpi": 0,
             "lon": 1,
             "lat": 1,
         },
@@ -669,7 +673,7 @@ def testdata_known_results():
             "mse_bias": 1,
             "mse_var": 0,
             "urmsd": 0,
-            "gpi": 1,
+            "gpi": 0,
             "lon": 1,
             "lat": 1,
         },
@@ -722,19 +726,19 @@ def testdata_random():
     expected = {
         (("col1_name", "col1"), ("reference_name", "reference")): {
             "n_obs": n - 2,
-            "gpi": 1,
+            "gpi": 0,
             "lon": 1,
             "lat": 1,
         },
         (("col2_name", "col2"), ("reference_name", "reference")): {
             "n_obs": n - 2,
-            "gpi": 1,
+            "gpi": 0,
             "lon": 1,
             "lat": 1,
         },
         (("reference_name", "reference"), ("zcol3_name", "zcol3")): {
             "n_obs": n - 2,
-            "gpi": 1,
+            "gpi": 0,
             "lon": 1,
             "lat": 1,
         },
@@ -783,7 +787,7 @@ def test_PairwiseIntercomparisonMetrics(testdata_generator, seas_metrics):
         metrics_calculators={(4, 2): (metrics_calculator.calc_metrics)},
     )
     results_pw = val.calc(
-        [1], [1], [1], rename_cols=False, only_with_reference=True
+        [0], [1], [1], rename_cols=False, only_with_reference=True
     )
 
     # in results_pw, there are four entries with keys (("c1name", "c1"),
@@ -852,7 +856,7 @@ def test_PairwiseIntercomparisonMetrics(testdata_generator, seas_metrics):
 
     print("running old setup")
 
-    results = val.calc(1, 1, 1, rename_cols=False)
+    results = val.calc(0, 1, 1, rename_cols=False)
 
     # results is a dictionary with one entry and key
     # (('c1name', 'c1'), ('c2name', 'c2'), ('c3name', 'c3'), ('refname',
@@ -940,7 +944,7 @@ def test_PairwiseIntercomparisonMetrics_confidence_intervals():
         },
     )
     results_pw = val.calc(
-        [1], [1], [1], rename_cols=False, only_with_reference=True
+        [0], [1], [1], rename_cols=False, only_with_reference=True
     )
 
     metrics_with_ci = {
@@ -961,7 +965,7 @@ def test_PairwiseIntercomparisonMetrics_confidence_intervals():
     # reconstruct dataframe
     frames = []
     for key in datasets:
-        frames.append(datasets[key]["class"].data)
+        frames.append(datasets[key]["class"].data[0])
     data = pd.concat(frames, axis=1)
     data.dropna(how="any", inplace=True)
 
@@ -1024,7 +1028,7 @@ def test_TripleCollocationMetrics(testdata_generator, seas_metrics):
         metrics_calculators={(4, 3): triplet_metrics_calculator.calc_metrics},
     )
     results_triplet = val_triplet.calc(
-        [1], [1], [1], rename_cols=False, only_with_reference=True
+        [0], [1], [1], rename_cols=False, only_with_reference=True
     )
 
     if "col1_name" in datasets.keys():
@@ -1088,7 +1092,7 @@ def test_TripleCollocationMetrics(testdata_generator, seas_metrics):
             },
         )
         results_triplet = val_triplet.calc(
-            [1], [1], [1], rename_cols=False, only_with_reference=True
+            [0], [1], [1], rename_cols=False, only_with_reference=True
         )
         for key in results_triplet:
             for dset, _ in key:
@@ -1152,7 +1156,7 @@ def test_temporal_matching_ascat_ismn():
         },
     )
     new_results = new_val.calc(
-        1, 1, 1, rename_cols=False, only_with_reference=True
+        0, 1, 1, rename_cols=False, only_with_reference=True
     )
 
     # old setup
@@ -1171,13 +1175,27 @@ def test_temporal_matching_ascat_ismn():
         temporal_matcher=old_matcher,
         metrics_calculators={(2, 2): metrics.calc_metrics},
     )
-    old_results = old_val.calc(1, 1, 1, rename_cols=False)
+    old_results = old_val.calc(0, 1, 1, rename_cols=False)
 
     old_key = (("ASCAT", "sm"), ("ISMN", "soil_moisture"))
     new_key = (("ASCAT", "sm"), ("ISMN", "soil_moisture"))
 
     assert old_results[old_key]["n_obs"] == new_results[new_key]["n_obs"]
 
+
+def test_TripleCollocationMetrics_failure():
+    """
+    Test if TripleCollocationMetrics returns the correct status attribute if
+    there is not enough data for bootstrapping.
+    """
+    df = make_some_data()
+    data = df.iloc[0:50][["ref", "k1", "k2"]]
+
+    triplet_metrics_calculator = TripleCollocationMetrics(
+        "ref", bootstrap_cis=True
+    )
+    res = triplet_metrics_calculator.calc_metrics(data, gpi_info=(0, 0, 0))
+    assert res["status"] == eh.METRICS_CALCULATION_FAILED
 
 # def test_sorting_issue():
 # GH #220