+++ tests/test_validation_framework/test_validation.py
@@ -40,6 +40,7 @@ import pytest
 import tempfile
 import warnings
 import pandas as pd
+import pytest
 
 import pygeogrids.grids as grids
 from pygeobase.io_base import GriddedTsBase
@@ -51,6 +52,7 @@ from pytesmo.validation_framework.data_manager import DataManager
 from pytesmo.validation_framework.results_manager import PointDataResults
 from pytesmo.validation_framework.validation import Validation
 from pytesmo.validation_framework.validation import args_to_iterable
+import pytesmo.validation_framework.error_handling as eh
 
 from pytesmo.validation_framework.metric_calculators import (
     PairwiseIntercomparisonMetrics)
@@ -62,6 +64,8 @@ from ismn.interface import ISMN_Interface
 with warnings.catch_warnings():
     warnings.filterwarnings("ignore")
     from ascat.read_native.cdr import AscatGriddedNcTs
+
+from .utils import create_datasets
 
 if __name__ != "__main__":
     from tests.test_validation_framework.test_datasets import (
@@ -109,8 +113,8 @@ def ascat_reader():
         "TUW_METOP_ASCAT_WARP55R22_{:04d}",
         grid_filename=grid_fname,
         static_layer_path=static_layers_folder,
+        ioclass_kws={'read_bulk': True}
     )
-    ascat_reader.read_bulk = True
     return ascat_reader
 
 
@@ -147,7 +151,7 @@ def check_results(
         vars_should = [
             u"n_obs", u"tau", u"gpi", u"RMSD", u"lon", u"p_tau", u"BIAS",
             u"p_rho", u"rho", u"lat", u"R", u"p_R", u"time", u"idx",
-            u"_row_size"
+            u"_row_size", u"status"
         ]
 
     with nc.Dataset(filename, mode="r") as results:
@@ -180,12 +184,6 @@ def test_ascat_ismn_validation(ascat_reader, ismn_reader):
         metadata = ismn_reader.read_metadata(idx)
         jobs.append((idx, metadata["longitude"].val, metadata["latitude"].val))
 
-    # Create the variable ***save_path*** which is a string representing the
-    # path where the results will be saved. **DO NOT CHANGE** the name
-    # ***save_path*** because it will be searched during the parallel
-    # processing!
-
-    save_path = tempfile.mkdtemp()
 
     # Create the validation object.
 
@@ -224,57 +222,53 @@ def test_ascat_ismn_validation(ascat_reader, ismn_reader):
         period=period,
     )
 
-    for job in jobs:
-        results = process.calc(*job)
-        netcdf_results_manager(results, save_path)
+    # Create the variable ***save_path*** which is a string representing the
+    # path where the results will be saved. **DO NOT CHANGE** the name
+    # ***save_path*** because it will be searched during the parallel
+    # processing!
+    #
 
-    results_fname = os.path.join(save_path,
-                                 "ASCAT.sm_with_ISMN.soil_moisture.nc")
-    # targets
-    target_vars = {
-        "n_obs": [
-            357,
-            384,
-            1646,
-            1875,
-            1915,
-            467,
-            141,
-            251
-        ],
-        # reference values for rho were inferred from the unscaled rho, since
-        # the CDF scaling should not change rank correlation
-        "rho": np.array([
-            0.5393457276100486,
-            0.7002289363385688,
-            0.6220023416557768,
-            0.5364715553916934,
-            0.3041366641200411,
-            0.6759989589212635,
-            0.8418980829611084,
-            0.7391369789045885,
-        ], dtype=np.float32),
-        # reference values for RMSD were taken from the first run with the new
-        # CDF matching implementation
-        "RMSD": np.array([
-            11.577185,
-            7.788051,
-            17.73936,
-            21.145535,
-            14.092733,
-            13.741866,
-            12.1061,
-            13.058796,
-        ], dtype=np.float32)
+    with tempfile.TemporaryDirectory() as save_path:
+        for job in jobs:
+            results = process.calc(*job)
+            netcdf_results_manager(results, save_path)
 
-    }
+        results_fname = os.path.join(save_path,
+                                     "ASCAT.sm_with_ISMN.soil_moisture.nc")
+        # targets
+        target_vars = {
+            "n_obs": [357, 384, 1646, 1875, 1915, 467, 141, 251],
+            "rho":
+                np.array([
+                    0.5393457276100486,
+                    0.7002289363385688,
+                    0.6220023416557768,
+                    0.5364715553916934,
+                    0.3041366641200411,
+                    0.6759989589212635,
+                    0.8418980829611084,
+                    0.7391369789045885,
+                ], dtype=np.float32),
+            "RMSD":
+                np.array([
+                    11.577185,
+                    7.788051,
+                    17.73936,
+                    21.145535,
+                    14.092733,
+                    13.741866,
+                    12.1061,
+                    13.058796,
+                ], dtype=np.float32)
+        }
 
-    check_results(
-        filename=results_fname,
-        target_vars=target_vars,
-    )
+        check_results(
+            filename=results_fname,
+            target_vars=target_vars,
+        )
 
-    ascat_reader.close()
+        ascat_reader.close()
+        del ascat_reader
 
 
 @pytest.mark.slow
@@ -406,7 +400,7 @@ def test_ascat_ismn_validation_metadata(ascat_reader, ismn_reader):
     vars_should = [
         'BIAS', 'R', 'RMSD', '_row_size', 'climate', 'gpi', 'idx', 'landcover',
         'lat', 'lon', 'n_obs', 'network', 'p_R', 'p_rho', 'p_tau', 'rho',
-        'station', 'tau', 'time'
+        'station', 'tau', 'time', 'status'
     ]
 
     check_results(
@@ -562,7 +556,23 @@ def test_validation_error_n2_k2():
 
 
 def test_validation_n3_k2_temporal_matching_no_matches():
-    tst_results = {}
+    tst_results = {
+        (("DS1", "x"), ("DS2", "y")): {
+            'gpi': np.array([4], dtype=np.int32),
+            'lon': np.array([4.]),
+            'lat': np.array([4.]),
+            'status': np.array([eh.NO_TEMP_MATCHED_DATA], dtype=np.int32),
+            'n_obs': np.array([0], dtype=np.int32),
+            'R': np.array([np.nan], dtype=np.float32),
+            'p_R': np.array([np.nan], dtype=np.float32),
+            'rho': np.array([np.nan], dtype=np.float32),
+            'p_rho': np.array([np.nan], dtype=np.float32),
+            'RMSD': np.array([np.nan], dtype=np.float32),
+            'BIAS': np.array([np.nan], dtype=np.float32),
+            'tau': np.array([np.nan], dtype=np.float32),
+            'p_tau': np.array([np.nan], dtype=np.float32)
+        }
+    }
 
     datasets = setup_two_without_overlap()
 
@@ -586,7 +596,12 @@ def test_validation_n3_k2_temporal_matching_no_matches():
 
     jobs = process.get_processing_jobs()
     for job in jobs:
-        results = process.calc(*job)
+        with pytest.raises(eh.NoTempMatchedDataError):
+            results = process.calc(*job, handle_errors="raise")
+
+    jobs = process.get_processing_jobs()
+    for job in jobs:
+        results = process.calc(*job, handle_errors="ignore")
         assert sorted(list(results)) == sorted(list(tst_results))
 
 
@@ -813,21 +828,26 @@ def test_validation_n3_k2():
 
 
 def test_validation_n3_k2_temporal_matching_no_matches2():
+
+    empty_result = {
+        'gpi': np.array([4], dtype=np.int32),
+        'lon': np.array([4.]),
+        'lat': np.array([4.]),
+        'status': np.array([eh.NO_TEMP_MATCHED_DATA], dtype=np.int32),
+        'n_obs': np.array([0], dtype=np.int32),
+        'R': np.array([np.nan], dtype=np.float32),
+        'p_R': np.array([np.nan], dtype=np.float32),
+        'rho': np.array([np.nan], dtype=np.float32),
+        'p_rho': np.array([np.nan], dtype=np.float32),
+        'RMSD': np.array([np.nan], dtype=np.float32),
+        'BIAS': np.array([np.nan], dtype=np.float32),
+        'tau': np.array([np.nan], dtype=np.float32),
+        'p_tau': np.array([np.nan], dtype=np.float32)
+    }
+
+    # only DS1 and DS3 overlap, so we only get two valid results
     tst_results = {
-        (("DS1", "x"), ("DS3", "y")): {
-            "n_obs": np.array([1000], dtype=np.int32),
-            "tau": np.array([np.nan], dtype=np.float32),
-            "gpi": np.array([4], dtype=np.int32),
-            "RMSD": np.array([0.0], dtype=np.float32),
-            "lon": np.array([4.0]),
-            "p_tau": np.array([np.nan], dtype=np.float32),
-            "BIAS": np.array([0.0], dtype=np.float32),
-            "p_rho": np.array([0.0], dtype=np.float32),
-            "rho": np.array([1.0], dtype=np.float32),
-            "lat": np.array([4.0]),
-            "R": np.array([1.0], dtype=np.float32),
-            "p_R": np.array([0.0], dtype=np.float32),
-        },
+        (("DS1", "x"), ("DS2", "y")): empty_result,
         (("DS1", "x"), ("DS3", "x")): {
             "n_obs": np.array([1000], dtype=np.int32),
             "tau": np.array([np.nan], dtype=np.float32),
@@ -842,6 +862,22 @@ def test_validation_n3_k2_temporal_matching_no_matches2():
             "R": np.array([1.0], dtype=np.float32),
             "p_R": np.array([0.0], dtype=np.float32),
         },
+        (("DS1", "x"), ("DS3", "y")): {
+            "n_obs": np.array([1000], dtype=np.int32),
+            "tau": np.array([np.nan], dtype=np.float32),
+            "gpi": np.array([4], dtype=np.int32),
+            "RMSD": np.array([0.0], dtype=np.float32),
+            "lon": np.array([4.0]),
+            "p_tau": np.array([np.nan], dtype=np.float32),
+            "BIAS": np.array([0.0], dtype=np.float32),
+            "p_rho": np.array([0.0], dtype=np.float32),
+            "rho": np.array([1.0], dtype=np.float32),
+            "lat": np.array([4.0]),
+            "R": np.array([1.0], dtype=np.float32),
+            "p_R": np.array([0.0], dtype=np.float32),
+        },
+        (("DS2", "y"), ("DS3", "x")): empty_result,
+        (("DS2", "y"), ("DS3", "y")): empty_result,
     }
 
     datasets = setup_three_with_two_overlapping()
@@ -865,8 +901,13 @@ def test_validation_n3_k2_temporal_matching_no_matches2():
 
     jobs = process.get_processing_jobs()
     for job in jobs:
-        results = process.calc(*job)
+        results = process.calc(*job, handle_errors="ignore")
         assert sorted(list(results)) == sorted(list(tst_results))
+
+    jobs = process.get_processing_jobs()
+    for job in jobs:
+        with pytest.raises(eh.NoTempMatchedDataError):
+            results = process.calc(*job, handle_errors="raise")
 
 
 def test_validation_n3_k2_masking_no_data_remains():
@@ -1152,7 +1193,7 @@ def test_ascat_ismn_validation_metadata_rolling(ascat_reader, ismn_reader):
     }
     vars_should = [
         u"gpi", u"RMSD", u"lon", u"lat", u"R", u"p_R", u"time", u"idx",
-        u"_row_size"
+        u"_row_size", u"status"
     ]
     for key, value in metadata_dict_template.items():
         vars_should.append(key)
@@ -1235,94 +1276,6 @@ def test_args_to_iterable_mixed_strings():
 #######################################################################
 # Tests for new temporal matcher & metric calculators
 #######################################################################
-
-
-def create_correlated_data(n_datasets, n, r):
-    """Creates n_datasets random timeseries with specified correlation"""
-    C = np.ones((n_datasets, n_datasets)) * r
-    for i in range(n_datasets):
-        C[i, i] = 1
-    A = np.linalg.cholesky(C)
-
-    return (A @ np.random.randn(n_datasets, n)).T
-
-
-class DummyReader:
-
-    def __init__(self, dfs, name):
-        self.data = [pd.DataFrame(dfs[i][name]) for i in range(len(dfs))]
-
-    def read(self, gpi, *args, **kwargs):
-        return self.data[gpi]
-
-
-class DummyNoneReader:
-
-    def __init__(self, dfs, name):
-        self.data = [pd.DataFrame(dfs[i][name]) for i in range(len(dfs))]
-
-    def read(self, gpi, *args, **kwargs):
-        names = self.data[gpi].columns
-        return pd.DataFrame(np.zeros((0, len(names))), columns=names)
-
-
-def create_datasets(n_datasets, npoints, nsamples, missing=False):
-    """
-    Creates three datasets with given number of points to compare, each
-    having number of samples given
-    """
-    dfs = []
-    for gpi in range(npoints):
-        r = np.random.rand()
-        data = create_correlated_data(n_datasets, nsamples, r)
-        index = pd.date_range("1980", periods=nsamples, freq="D")
-        dfs.append(
-            pd.DataFrame(
-                data,
-                index=index,
-                columns=(["refcol"] +
-                         [f"other{i}col" for i in range(1, n_datasets)])))
-
-    datasets = {}
-    datasets["0-ERA5"] = {
-        "columns": ["refcol"],
-        "class": DummyReader(dfs, "refcol")
-    }
-    for i in range(1, n_datasets - 1):
-        datasets[f"{i}-ESA_CCI_SM_combined"] = {
-            "columns": [f"other{i}col"],
-            "class": DummyReader(dfs, f"other{i}col")
-        }
-    if missing:
-        datasets[f"{n_datasets - 1}-missing"] = {
-            "columns": [f"other{n_datasets - 1}col"],
-            "class": DummyNoneReader(dfs, f"other{n_datasets - 1}col")
-        }
-    else:
-        datasets[f"{n_datasets - 1}-ESA_CCI_SM_combined"] = {
-            "columns": [f"other{n_datasets - 1}col"],
-            "class": DummyReader(dfs, f"other{n_datasets - 1}col")
-        }
-    return datasets
-
-
-def test_missing_data():
-    n_datasets = 5
-    npoints = 5
-    nsamples = 100
-
-    datasets = create_datasets(n_datasets, npoints, nsamples, missing=True)
-
-    metric_calculator = PairwiseIntercomparisonMetrics()
-
-    val = Validation(
-        datasets,
-        spatial_ref="0-ERA5",
-        metrics_calculators={(n_datasets, 2): metric_calculator.calc_metrics},
-        temporal_matcher=make_combined_temporal_matcher(pd.Timedelta(12, "H")),
-    )
-    gpis = list(range(npoints))
-    val.calc(gpis, gpis, gpis, rename_cols=False, only_with_reference=True)
 
 
 def test_combined_matching_scaling():